payments/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
blinker==1.7.0
Brlapi==0.8.3
certifi==2020.6.20
cfgv==3.3.1
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.3
django-filter==24.3
djangorestframework==3.15.1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
Flask==3.0.2
frappe==0.0.1
frappe-bench==5.22.3
frozenlist==1.4.1
future==0.18.2
geoip2==4.8.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
honcho==1.1.0
httplib2==0.20.2
identify==2.4.10
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
maxminddb==2.6.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
nodeenv==0.13.4
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==11.9.0
platformdirs==2.5.1
pre-commit==2.17.0
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodome==3.20.0
pycryptodomex==3.20.0
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-crontab==2.6.0
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzipper==0.3.6
reportlab==3.6.8
requests==2.25.1
sdkmanager==0.5.1
SecretStorage==3.3.1
semantic-version==2.8.5
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.28
sqlparse==0.4.4
supervisor==4.2.1
systemd-python==234
toml==0.10.2
tomli==2.0.1
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
